Fixed integer overflow with CAS and doc sizes.

+++ src/main/java/gehring/uima/distributed/benchmark/BenchmarkMetadataProvider.java
@@ -7,8 +7,8 @@ public interface BenchmarkMetadataProvider {
 
 	public int getNumberOfDocuments();
 
-	public int getSumOfAllCasSizes();
+	public long getSumOfAllCasSizes();
 
-	public int getSumOfAllDocumentSizes();
+	public long getSumOfAllDocumentSizes();
 
 }
